# PodSecurity: benchmark and optimize privileged namespace evaluations