--- a/staging/src/k8s.io/pod-security-admission/admission/main_test.go
+++ b/staging/src/k8s.io/pod-security-admission/admission/main_test.go
@@ -1,0 +1,36 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package admission
+
+import (
+	"fmt"
+	"os"
+	"reflect"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	sharedCopy := sharedAllowedResponse().DeepCopy()
+	rc := m.Run()
+
+	if !reflect.DeepEqual(sharedCopy, sharedAllowedResponse()) {
+		fmt.Println("sharedAllowedReponse mutated")
+		rc = 1
+	}
+
+	os.Exit(rc)
+}