… plugin/pkg/admission/security/podsecurity/testdata/pod.yaml