#### What type of PR is this?

/kind feature

Add benchmarks and optimize the "all privileged" case (which is the default for clusters upgraded to a version that enables this admission plugin)

```
benchmark                                                           old ns/op     new ns/op     delta
BenchmarkVerifyPod/enforce-implicit_pod-12                          1595          225           -85.89%
BenchmarkVerifyPod/enforce-implicit_deployment-12                   1703          234           -86.28%
BenchmarkVerifyPod/enforce-privileged_pod-12                        1600          245           -84.68%
BenchmarkVerifyPod/enforce-privileged_deployment-12                 1712          254           -85.18%
BenchmarkVerifyPod/enforce-baseline_pod-12                          2906          2850          -1.93%
BenchmarkVerifyPod/enforce-baseline_deployment-12                   1714          255           -85.14%
BenchmarkVerifyPod/enforce-restricted_pod-12                        3170          3125          -1.42%
BenchmarkVerifyPod/enforce-restricted_deployment-12                 1726          261           -84.88%
BenchmarkVerifyPod/warn-baseline_pod-12                             2883          2841          -1.46%
BenchmarkVerifyPod/warn-baseline_deployment-12                      2985          2913          -2.41%
BenchmarkVerifyPod/warn-restricted_pod-12                           3143          3176          +1.05%
BenchmarkVerifyPod/warn-restricted_deployment-12                    3295          3157          -4.19%
BenchmarkVerifyPod/enforce-warn-audit-baseline_pod-12               5210          5132          -1.50%
BenchmarkVerifyPod/enforce-warn-audit-baseline_deployment-12        4213          4069          -3.42%
BenchmarkVerifyPod/warn-baseline-audit-restricted_pod-12            4387          4252          -3.08%
BenchmarkVerifyPod/warn-baseline-audit-restricted_deployment-12     4415          4316          -2.24%

benchmark                                                           old allocs     new allocs     delta
BenchmarkVerifyPod/enforce-implicit_pod-12                          12             1              -91.67%
BenchmarkVerifyPod/enforce-implicit_deployment-12                   14             1              -92.86%
BenchmarkVerifyPod/enforce-privileged_pod-12                        12             1              -91.67%
BenchmarkVerifyPod/enforce-privileged_deployment-12                 14             1              -92.86%
BenchmarkVerifyPod/enforce-baseline_pod-12                          17             17             +0.00%
BenchmarkVerifyPod/enforce-baseline_deployment-12                   14             1              -92.86%
BenchmarkVerifyPod/enforce-restricted_pod-12                        17             17             +0.00%
BenchmarkVerifyPod/enforce-restricted_deployment-12                 14             1              -92.86%
BenchmarkVerifyPod/warn-baseline_pod-12                             17             17             +0.00%
BenchmarkVerifyPod/warn-baseline_deployment-12                      19             19             +0.00%
BenchmarkVerifyPod/warn-restricted_pod-12                           17             17             +0.00%
BenchmarkVerifyPod/warn-restricted_deployment-12                    19             19             +0.00%
BenchmarkVerifyPod/enforce-warn-audit-baseline_pod-12               27             27             +0.00%
BenchmarkVerifyPod/enforce-warn-audit-baseline_deployment-12        24             24             +0.00%
BenchmarkVerifyPod/warn-baseline-audit-restricted_pod-12            22             22             +0.00%
BenchmarkVerifyPod/warn-baseline-audit-restricted_deployment-12     24             24             +0.00%

benchmark                                                           old bytes     new bytes     delta
BenchmarkVerifyPod/enforce-implicit_pod-12                          2120          112           -94.72%
BenchmarkVerifyPod/enforce-implicit_deployment-12                   2304          112           -95.14%
BenchmarkVerifyPod/enforce-privileged_pod-12                        2120          112           -94.72%
BenchmarkVerifyPod/enforce-privileged_deployment-12                 2304          112           -95.14%
BenchmarkVerifyPod/enforce-baseline_pod-12                          3368          3368          +0.00%
BenchmarkVerifyPod/enforce-baseline_deployment-12                   2304          112           -95.14%
BenchmarkVerifyPod/enforce-restricted_pod-12                        3368          3368          +0.00%
BenchmarkVerifyPod/enforce-restricted_deployment-12                 2304          112           -95.14%
BenchmarkVerifyPod/warn-baseline_pod-12                             3368          3368          +0.00%
BenchmarkVerifyPod/warn-baseline_deployment-12                      3552          3552          +0.00%
BenchmarkVerifyPod/warn-restricted_pod-12                           3368          3368          +0.00%
BenchmarkVerifyPod/warn-restricted_deployment-12                    3552          3552          +0.00%
BenchmarkVerifyPod/enforce-warn-audit-baseline_pod-12               5864          5864          +0.00%
BenchmarkVerifyPod/enforce-warn-audit-baseline_deployment-12        4800          4800          +0.00%
BenchmarkVerifyPod/warn-baseline-audit-restricted_pod-12            4616          4616          +0.00%
BenchmarkVerifyPod/warn-baseline-audit-restricted_deployment-12     4800          4800          +0.00%
```

#### Does this PR introduce a user-facing change?
```release-note
NONE
```

xref https://github.com/kubernetes/enhancements/issues/2579
/cc @tallclair