@liggitt 2021-09-20
this is ready for review, and should go in before we start making changes for metrics and aggregation