@liggitt 2021-09-21
comments addressed