@liggitt 2021-09-21
weird:
```
ERROR: failed to create cluster: docker run error: command "docker run --name kind-control-plane --hostname kind-control-plane --label io.x-k8s.kind.role=control-plane --privileged --security-opt seccomp=unconfined --security-opt apparmor=unconfined --tmpfs /tmp --tmpfs /run --volume /var --volume /lib/modules:/lib/modules:ro -e KIND_EXPERIMENTAL_CONTAINERD_SNAPSHOTTER --device /dev/fuse --detach --tty --label io.x-k8s.kind.cluster=kind --net kind --restart=on-failure:1 --init=false --publish=127.0.0.1:45105:6443/TCP -e KUBECONFIG=/etc/kubernetes/admin.conf kindest/node:latest" failed with error: exit status 125
```

/retest